+++ lsp-server/diplomat/src/diplomat_lsp.cpp
@@ -86,12 +86,16 @@ _project_file_tree_valid(false)
     ServerCapabilities_workspace sc_ws;
     sc_ws.workspaceFolders = ws;
 
+    CompletionOptions sc_completion;
+    sc_completion.resolveProvider = false;
+
     capabilities.textDocumentSync = sync;
     capabilities.workspace = sc_ws;
     capabilities.definitionProvider = true;
     capabilities.referencesProvider = true;
     capabilities.documentFormattingProvider = true; // Can handle formatting options
     capabilities.renameProvider = true;
+    capabilities.completionProvider = sc_completion;
 
     _bind_methods();    
 }
@@ -165,6 +169,7 @@ void DiplomatLSP::_bind_methods()
 
     bind_notification("textDocument/didOpen", LSP_MEMBER_BIND(DiplomatLSP, _h_didOpenTextDocument));
     bind_notification("textDocument/didSave", LSP_MEMBER_BIND(DiplomatLSP, _h_didSaveTextDocument));
+    bind_request("textDocument/completion", LSP_MEMBER_BIND(DiplomatLSP, _h_completion));
     bind_request("textDocument/definition", LSP_MEMBER_BIND(DiplomatLSP, _h_gotoDefinition));
     bind_request("textDocument/formatting", LSP_MEMBER_BIND(DiplomatLSP, _h_formatting));
     bind_request("textDocument/references", LSP_MEMBER_BIND(DiplomatLSP, _h_references));